response codes
I think the basic server is almost done, now remains the client to be finished so i can quickly ship this app. Well technically i did ship this app earlier in the year so yeah you could say this is a  revised version of it. I think if i really focus i could finish it in one day but i am really struggling to find the motivation but oh well lets see how it goes